requirements.txt
streamlit
pandas
google-cloud-firestore
google-cloud-storage